requirements.txt
asgiref
Django==5.0.9
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
django-environ
djangorestframework
django-cors-headers